Cambios back

+++ back/src/main/java/tingeso/back/services/CalculateService.java
@@ -10,36 +10,36 @@ public class CalculateService {
     // r es la tasa de interes mensual
     // n es el numero total de pagos // plazo / 12 / 100
 
-    public float calculatePayment(float p, float interest, float term){
-        float result;
-        float r = interest / 100;
-        float n = term * 12;
+    public double calculatePayment(double p, double interest, double term){
+        double result;
+        double r = interest / 100;
+        double n = term * 12;
 
-        float base = (float) Math.pow((1 + r), n);
+        double base =  Math.pow((1 + r), n);
 
         result = p * ((r * base) / (base - 1));
 
         return result;
     }
 
-    public float calculateInsurance(float creditAmount, float n){
-        float fireInsurance = (float) 20.000;
-        float desgravamentInsurance = creditAmount * (float) 0.0003;
-        float desgravamentInsuranceMonthly = desgravamentInsurance * (n*12);
+    public double calculateInsurance(double creditAmount, double n){
+        double fireInsurance =  20.000;
+        double desgravamentInsurance = creditAmount *  0.0003;
+        double desgravamentInsuranceMonthly = desgravamentInsurance * (n*12);
         return fireInsurance + desgravamentInsuranceMonthly;
     }
 
-    public float calculateAdministrationCost(float creditAmount){
-        return creditAmount * (float) 0.01;
+    public double calculateAdministrationCost(double creditAmount){
+        return creditAmount *  0.01;
     }
 
-    public float calculateTotalCost(float M,
-                                    float n,
-                                    float insurance,
-                                    float administrationCost){
-        float monthlyFee = M + insurance;
+    public double calculateTotalCost(double M,
+                                     double n,
+                                     double insurance,
+                                     double administrationCost){
+        double monthlyFee = M + insurance;
 
-        float toMonthlyFee = n * 12;
+        double toMonthlyFee = n * 12;
 
         return (monthlyFee * toMonthlyFee) + administrationCost;
     }
